requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2.2
django-fontawesome-5==1.0.16
django-recaptcha2==1.4.1
idna==2.8
Pillow==6.0.0
pkg-resources==0.0.0
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
